remove test main for now
